+++ beacon-chain/core/epoch/precompute/BUILD.bazel
@@ -3,7 +3,8 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "processing.go",
+        "attestation.go",
+        "new.go",
         "type.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/epoch/precompute",
@@ -19,7 +20,10 @@ go_library(
 
 go_test(
     name = "go_default_test",
-    srcs = ["processing_test.go"],
+    srcs = [
+        "attestation_test.go",
+        "new_test.go",
+    ],
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/core/helpers:go_default_library",
